requirements.txt
pypdf==3.3.0
pillow==9.4.0
pyzbar==0.1.9